crawler/requirements.txt
redis
gitpython
tqdm